[perf] Cache level pattern values and period facts
ActivityFormulaVariables have a tendency to explode in numbers,
suppose a zone has 11 org units, the number of variables that need to
be generated soon reaches over 10 000. The faster we can get the inner
loop, the faster the whole process is.

This fixes the dates_substitution and the level pattern values from
being needed to be calculated each time.