Use concat to fix stack level too deep error
This was a fun one, in production we ran into an issue where running a
particular zone we would get a:

      SystemStackError: stack level too deep

And the backtrace pointed to the `push` line.

Turns out, that ruby has a hard limit on the number of arguments you
can give to a function (which is what we do when calling
`push(*arr)`), so because this particular zone was generating 300 000
variables, the `push` broke.

Using concat fixes this (and should make memory behave nicer as well).

I tried to figure out what the hard limit is:

      RubyVM::DEFAULT_PARAMS
      # => {
        :thread_vm_stack_size=>1048576,
        :thread_machine_stack_size=>1048576,
        :fiber_vm_stack_size=>131072,
        :fiber_machine_stack_size=>524288
      }

It seems to be the `fiber_vm_stack_size`, you can try this at home
with:

      [].push(*Array.new(131072))

Which will fail, but since you're already somewhere in the stack
you'll need to detuct a few hundred to get this to pass.