Fix: Analytics call ran multiple times for same period
If you had multiple packages, which had formulas which returned a
similar set of periods, the overlapping periods would result in a
double call.

This only happened if you had more than 5 periods (the point at which
we split the analytics call into multiple calls), so even if you had
overlapping ones, if the total was not over 5, it would work.

The reason the test didn't catch this was that all the
package_arguments were set up with the same array of periods, so when
we ran:

    map(&:periods).uniq

It compared the arrays, instead of the elements, and because the
arrays were exactly the same, we had no problems.

I've made the spec a bit more wordy about this, and set up a scenario
where we are triggering this behavior.

I've also sorted the resulting periods, mostly because this reads
nicer when debugging the calls. (And much easier to stub the requests)